requirements.txt
# Core dependencies
watchdog==3.0.0
openai==1.12.0
boto3==1.34.34
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
Pillow==10.2.0
httpx==0.26.0

# Development dependencies (optional)
pytest==7.4.3
black==23.11.0
flake8==6.1.0 